requirements.txt
# Core dependencies for SPRING Hybrid Training System
# Python 3.8+ required

# Deep Learning Framework
torch==2.4.1
torchvision>=0.19.0

# Numerical Computing
numpy==1.23.0
scipy==1.10.1

# Computer Vision & Image Processing
opencv-python==4.11.0
Pillow==10.2.0

# Hugging Face Libraries for DETR and Stable Diffusion
transformers==4.39.3
diffusers==0.27.2
accelerate>=0.20.0

# Visualization & Plotting
matplotlib==3.5.2

# Progress Bars
tqdm==4.63.1

# Logging & Monitoring (optional but recommended)
tensorboard>=2.10.0
wandb>=0.15.0

# Data Processing
jsonlines>=3.1.0

# CUDA support (optional, install if using GPU)
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/
# Example: pip install torch==2.4.1+cu121 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121